js: refactor parser into sub-folder
