+++ src/langs/javascript/index.ts
@@ -1,0 +1,4 @@
+export * from "./filter.ts";
+export * from "./generate.ts";
+export * from "./parser/index.ts";
+export * from "./types.ts";
